Inline vanity handler
